Reraise exceptions from harvest thread in the main thread. PYTHON-2158
When running tests that use the `collector_agent_registration_fixture`, if an
unhandled exception is raised in the harvest thread, reraise it in the main
thread. This way our tests can alert us if something goes wrong in the harvest
thread.

Additionally, if a custom metric is created that starts with
`'Supportability/Python/Harvest/Exception'` but is not a `DiscardDataForRequest`
or `RetryDataForRequest` error, then create an exception for that as well. This
is necessary because the harvest thread handles these errors.