Handle custom connection_pool in redis instance_info. PYTHON-2167
Our redis instrumentation gets the datastore instance info from
client.connection_pool.connnection_kwargs. But, that fails in the
case of Redis Blaster which defines it's own connection pool class
that does not have the connection_kwargs attribute.

We handle that case by not recording instance info for redis, if
instance_info fails.