Inline function calls in redis instrumentation. PYTHON-2167
Remove `_client_instance_info` and `_connection_instance_info` and
just call `_instance_info` directly. There's still a function call
needed to convert connection instance attributes to a dict, but the
`_getattr` helper function has been removed.

The purpose of this refactor is to save the overhead of unneeded
function calls.