Test multipart command for redis custom connection pool. PYTHON-2167
Add test with the 'CLIENT LIST' multipart command to verify that
we don't throw an error when the connection pool lacks the
connection_kwargs attribute.

To test that we don't throw an error in send_command when using a
connection pool without the connection_kwargs attribute, we create
a FakeConnectionPool and add that to the redis client.

Also, separate out Redis Blaster tests to verify that we record
instance info for single server commands (ones that go through
redis.Connection:send_command), but don't record instance info
for multi-server commands.