Disable wrapper metric check for cornice. PYTHON-2199
Reporting the cornice services "wrapper" function as a metric is
dependent on the version of cornice being used. In earlier versions of
cornice, the wrapper was implemented incorrectly and incorrectly changed
the __name__ and __module__ attributes.

In later versions of cornice, this issue has been corrected and the
wrapper funtion is no longer reported as a metric.