Update python-base to ubuntu:16.04. PYTHON-1868
Upgrades the ubuntu version of python-base image from 12.04 to 16.04 and
correctly installs packages required for testing.

Upgrade to oursql==0.9.3.2 because the MySQL version was also upgraded. The
earlier version of oursql would no longer work. See
https://github.com/python-oursql/oursql/pull/4

Add CFLAGS to umemcache which will compile the binaries and ignore any
warnings. Without these flags, warnings are considered errors and the package
cannot be installed.