parseconfig.py script work with new tox files. PYTHON-2207
The dep section of the new tox files will look something like:

```
[testenv]
deps =
    Flask-Compress
    flask0006: flask<0.7
    flask0007: flask<0.8
    flask0008: flask<0.9
    flask0009: flask<0.10
    flask0010: flask<0.11
```

The difference with the old style toxes being the colons.

The second big change that needs to be accounted for is the removal of the
`[testenv:py-...]` sections which we used for discovering which python versions
to use. With the removal of these sections, now all packages will be built for
all versions of python.

Uses a regular expression to discover which python versions to test against. We
will attempt to download the package for all python versions included in the
tox file whether or not the test will actually be run against it.

Also adds a unittest file and adds it to jenkins testing.