Fix the rare cache hit then removed failure. PYTHON-2212
With multiple tests running on the same machine, it's possible for the
build to detect a cache hit and then shortly after that it's removed by
another job that's running.

This would result in a build attempting to use a layer in the cache even
though the layer has already been removed. This results in a build
failure.