Use -f to remove docker containers during cleanup. PYTHON-2214
It's possible that packnsend can receive a SIGTERM (when the user
presses CTRL+C to stop a run for example) while the containers are
running.

In this case, the container can only be removed if the "-f" flag is
passed into docker rm.  Force should be the desired behavior since
failure to remove the container would result in the image and volumes
also failing to get removed.