Only delete paths that we have inserted. PYTHON-2055
The intent of the sys.path modification code in
bootstrap/sitecustomize.py is to bootstrap the newrelic agent path for
import in the case that it's not already in the PATH.

This allows the newrelic-admin script to import the agent when it's
installed by buildout. (newrelic-admin is in the bin directory but
doesn't have newrelic in the import path).

Before this code modification, the assumption is that the newrelic
library path is already excluded from the syspath.

With extra_path removed, it's possible that the library path is a global
directory like .../site-packages. In that case, we don't insert a path
and we consequently don't need to remove the path.