Use the original view func when creating callable_name. PYTHON-2248
The situation arises from the fact that django rest framework does not set
`__qualname__` on WrappedAPIView (see
https://github.com/tomchristie/django-rest-framework/blob/master/rest_framework/decorators.py#L57-L58)
whereas it sets `__name__` and `__module__` both from the original view
function.  Therefore, in Python 2 the transaction names were correct but in
Python 3 (which relies on the `__qualname__` attr) was getting named something
like "/publisher.api_v2_views:WrappedAPIView.handler".

The solution is to call `callable_name` on the original view function instead
of the created view handler.

Then, since there is already a metric with the name returned by
`callable_name(func)` when an app uses Python 2, we must go one step further
and that is to tack on the request method.