Make calls to the WSGI entrypoint self-consistent. PYTHON-2262
Today, we inject middleware when the browser agent needs to be injected
(browser monitoring enabled).

The middleware's entrypoint is through the __call__ function in
newrelic/api/web_transaction.py

Currently, calling __call__ immediately returns a generator which
results in the application's entrypoint FunctionTrace segment being
artificially small.

There are many philosophies for how WSGI/Response time is measured but
at a minimum, we should be self-consistent with how we call into our
customer's application.

This commit calls the customer's WSGI entrypoint before
creating/returning the generator which matches the behavior of the agent
without middleware injected.