Remove Content-Length tests from django. PYTHON-2262
This django regression test is to check that 'Content-Length' is not
inserted by our NR browser agent logic if 'Content-Length' is not
already there.

The problem is, with the modifications to the browser middleware
WebTest/webob (our WSGI server for testing) now always inserts
'Content-Length'.

The reason webob now inserts Content-Length is because of this logic:
https://github.com/Pylons/webob/blob/8669fe335b54697f23a787f67da19552bc03d5c6/webob/request.py#L1269-L1272

As a result, these tests have been moved into a unit test for the
browser middleware.