Instrument handle_uncaught_exception in Django 1.10. PYTHON-2268
A regression was created in Django 1.10 where the agent was no longer creating
error events for exceptions raised during template rendering. One such example
is when Django throws a `TemplateSyntaxError` due to a missing
`{% load static %}` statement in a template file.

This was happening because the `handle_uncaught_exception` method was moved in
Django 1.10. The solution is to instrument this method in its new location.