Implement stackless node parenting. PYTHON-2279
In async programs, it's possible that functions can schedule children
for execution and then immediately terminate.

An example of this model is the Python Tornado framework. Tornado
implements an event-loop which maintains a list of scheduled callbacks.
Each routine that's running returns control to the event loop which runs
callbacks. Under this model, scheduled callbacks can run after the
scheduler. In this case, the parent must defer its data finalization
phase until the children have recorded their respective times.

In addition, it's possible for 2 children to have overlapping start/end
times. Consider a threaded program. Each thread starts/ends
independently with the context switching managed outside the main
program. In order to support this case, we need to consider 2
concurrently running children with the same parent.