Disable motor instrumentation for version >= 0.6. PYTHON-1907
Motor >= 0.6 disallows access to attributes that start with a leading
underscore. This eliminates any incompatibilities with the python agent
and callable_name.

The background behind this issue is that callable_name caches the
object path as an attribute on the object. Previously, accessing that
attribute would cause a lookup of a database named after the cached
object path. This behavior was clearly undesired so we patched an
AttributeError when attempting to access new relic specific attributes.

This is no longer necessary since those New Relic attributes can't be
accessed in Motor >= 0.6 anyway (since they start with a leading
underscore).