Handle coroutine naming for Tornado 4.5. PYTHON-2329
Tornado 4.5 updated its runner implementation so that it no longer uses
lambdas for scheduling coroutine callbacks.

Prior to Tornado 4.5, the Python Agent used the return value of the
Runner run call to attach the coroutine name. Unfortunately, the return
value is no longer passed to the future in the new implementation.

In order to track naming in Tornado 4.5, the name is attached to the
Runner.run future and then to the callback when the future is scheduled.