Test internal metric is created only on exit. PYTHON-2342
This test is important because it's possible that internal metrics may
be sent up during the same harvest cycle that they are generated.

However, if there are internal metrics which haven't completed, they
should be reported at the next harvest cycle.

To test for the "next harvest cycle" behavior, we verify that
reset_metric_stats does not interfere with metrics which are waiting to
be generated.