Stop creation of agent instance from api call wrapper. PYTHON-2346
uwsgi (amongst other applications) allow parts of the agent code to
execute before forking. That means that we shouldn't allow the agent to
become activated solely from using APIs.

The agent should only activate the harvest thread when a transaction
starts. This is the only location that the agent finally knows where
transactions are going to run.