Test exception raised to django's error handler. PYTHON-2438
When settings.TASTYPIE_FULL_DEBUG is True, exceptions are re-raised by
the tastypie framework if a NotFound exception is hit. Therefore the
regular _handle_500 method is skipped. In this case, error handling is
done by django itself.

This commit adds tests for this. Because when caught by django's error
handler the response code is 500, in most cases, we will want to record
the error.