Fix non-streaming response gRPC future tests. PYTHON-2483
For non-streaming responses, there's a race condition on the iterator.
If the response thread returns prior to calling the iterator, the
iterator immediately sends a StopIteration. However, if the response
thread is running when the iterator is called, a list of messages will
be returned.

The way to handle this race condition is to try the iterator first
(handles the streaming response case) followed by a call to result().