Pin moto version in botocore tests. PYTHON-2517
Moto has a lot of issues right now. The one that's causing Jenkins
failures is an attempt to import boto3 from the aws polly mock
https://github.com/spulec/moto/blob/acb1cac0b7da8f5f18638caa629ff59455a999c2/moto/polly/models.py#L5

The boto3 library now imports from botocore and expects certain types of
exceptions to be present. These exceptions are not present in earlier
versions of botocore (such as the ones being tested) which results in an
ImportError.

Botocore tests should not import boto3 since doing so can import
botocore at a different version than expected.