Add response processing to aiohttp. PYTHON-2513
Unfortunately, AIOHTTP transforms exceptions to 500 responses in a
location that's outside the current scope of the transaction. Moving the
instrumentation point any farther up is not a great solution because we
would end up having to create separate instrumentation for each server
that's driving aiohttp applications.

Instead of moving the instrumentation, when a 500 error occurs we save
the CAT headers to the request object so that when the response is
processed, the CAT headers are attached.

An assumption is also made that exceptions raised in the throw/send
functions results in a 500 response (even though the actual response
codes are not captured).