Ignore cancelled aiohttp transaction. PYTHON-2531
In an async model, Futures can be cancelled at any yield point.
Cancellation can occur under normal network circumstances such as
whenever a client disconnects. In that case, the transaction may not
have reached its final name. Since the naming is not guaranteed (and the
transaction may be ignored downstream), reporting the transaction can
result in a MGI.

Data on these transactions shouldn't be too useful since the
cancellation is likely the result of the client disconnecting and the
full transaction never completes.