Reduce complexity of request handler wrapping. PYTHON-2549
The goal of this commit is to reduce some of the complexity in the way
we wrap request handlers during application initialization.

First, instead of calling `callable_name` on the method, just get the
name for the FunctionTrace from the transaction name since we know this
is already the correct value.

Second, instead of setting a `_nr_wrap_complete` flag on the handler,
just check to see if the method we want to wrap is an instance of
`_NRBoundFunctionWrapper`.