Use enabled flag to gate enter + drop logic in aiohttp. PYTHON-2573
Websockets can cause a transaction that would have otherwise been
enabled (since the _settings property is populated) to become disabled.

Since we don't currently record websocket requests as transactions, the
web_transaction logic sets the transaction enabled flag to false when it
detects a websocket upgrade request.

Prior to this fix, this causes a crash since the aiohttp instrumentation
assumes __enter__ will result in a transaction being populated in the
cache.