Remove cast to bool in should_record_params. PYTHON-2606
The result of should_record_params must be cast to a bool at some point.
This can happen explicitly or implicitly as part of evaluating
should_record_params in an if statement. Tracing through the CPython
core code, it appears to be slightly more efficient to avoid explicitly
casting to a bool.