Ensure finished futures are handled correctly. PYTHON-2576
`Future.add_done_callback` is changing with Tornado v5.0. From their
docs:

> The timing of callbacks scheduled with Future.add_done_callback has changed

The particular difference is that when the future is done, instead of
calling the callback directly, it is instead added to the ioloop for
next available execution. This means that the callback _could_ be called
later than it did in previous Tornado versions.

This becomes a problem when dealing with external traces that have been
popped. We must ensure that these traces are exited before the
transaction exits or else we get the 'Transaction ended but current_node
is not Sentinel' error.

The solution is to add our instrumentation to the `handle_response`
callback directly instead of using `add_done_callback`.