Add test to expose pika memory leak. PYTHON-2620
A customer has reported that New Relic causes references to channels to
be retained. This is ultimately caused by the approach that was taken
for 'no-trace' method filtering.

The bound method is added to a set in the instrumentation which is never
emptied. This results in all channel objects being retained for the
lifecycle of the python process since the bound method references the
channel object.

This test exposes the problem by creating unreferenced channels and
calling gc.collect(). The expected behavior is that the channel objects
are deleted.