Work around bug in Tornado 4.3 in our testing. PYTHON-2642
This was a bug in our tests for Tornado 4.3 only.

In Tornado 4.3, the HTTPHeaders class was introduced. Unfortunately, it
did not have a `__str__` method, so calling `str` on it would return
garbage like `'<tornado.httputil.HTTPHeaders object at 0x10da8af10>'`.

By specifying `__dict__` we can work around this bug for all Tornado
versions.