Re-enable use of the CustomHTTPClient. PYTHON-2641
Tornado allows customers to change the default on the AsyncHTTPClient
configurable. Our instrumentation supports wrapping custom fetch_impl
methods on configured HTTP clients.

This commit updates our tests to use the configure() method to change
the HTTPClient type for the purposes of our testing.

Custom HTTP clients that are configured should be supported and the test
checks for that behavior.