Add a finalizer for TraceContext. PYTHON-2651
It's possible that generators that aren't completely consumed will go
out of scope. In this case, the agent behavior should be to immediately
end and record the coroutine trace that was started / in progress.

The customer would expect that a generator that's been used but not
entirely consumed would be reported to New Relic.