Fix client cat tests for aiohttp v1.x. PYTHON-2678
This commit fixes the tests in framework_aiohttp/test_client_cat.py.

1. ClientSession does not have kwarg raise_for_status. For earlier
versions of the framework, manually raise the exception by calling
`response.raise_for_status()`.
2. The location of errors has changed.
3. Aiohttp has changed the location of where headers are written. In
version 2.x it is in the PayloadWriter class. But this module does not
exist in this earlier version. This will be addressed in a later commit.
4. Header casing is strangely enforced. This will be addressed in a
later commit.