Test flask compress with html served from files. PYTHON-2701
Flask has a very clever mechanism for reading files from disk and
serving them through a wsgi server.

This mechanism (send_file) opens a file handle, and returns an iterator
that, when iterated on, sends up to a certain number of bytes back to
the wsgi server. The iterator ends when the file has been completely
read.

This special iterator usage is signaled by setting the
direct_passthrough attribute on the response object. Setting this
attribute prevents any access to the contents of the response unless the
response is iterated through.

It's possible that the content-type header for the streamed file can be
text/html.

With Flask-Compress, if the file has a content type of text/html, it
will be gzip compressed. In order for Flask-Compress to read and
compress the file from the WSGI iterator, Flask-Compress sets the
direct_passthrough attribute back to False.

This test exposes the behavior of the agent in that case.