Add test asserting exclusive time. PYTHON-2686
This test asserts that the exclusive time is calculated correctly when
using ensure_future. If the parenting is wrong, as it was before the
bugfix PYTHON-2686, then the `_coro` function trace node will get
parented under Sentinel. This means that the exclusive time for the
`main` function will be too big.