Assign event loop to avoid traceback in py34. PYTHON-2779
Python 3.4 will throw an error here if we do not set the event loop.

> RuntimeError: There is no current event loop in thread 'MainThread'.

This is because `ensure_future` will attempt to get the current event
loop. But since the running event loop is in a different thread than the
target application, according to Python 3.4 it does not have automatic
access to this event loop. The avoid this, we set the loop as the
current loop in this thread.