Add distributed trace intrinsics to slow sql traces. PYTHON-2862
This commit moves the location where the `guid`, `sampled`, `priority`,
and `traceId` intrinsics are added. This moves all distributed tracing
related intrinsics to a single method but does not change any
functionality.

With this change we are then able to add all these distributed tracing
intrinsics to slow sql traces.