Optimize object allocations in create_workspace. PYTHON-2879
Using a copy.copy to create the workspace avoids running the __init__ of
StatsEngine and therefore avoids the allocation of all of the reservoir
and container objects.

The call to reset_stats causes new objects to be allocated maintaining
the isolation between the original stats object and the work area.