Prevent heapq from ever comparing incomparable entries. PYTHON-2964
In the case that two entries have the same priority, "sample" will be used as
the tiebreaker comparison by the heap. This will throw a TypeError in Python 3
if both entries' dicts aren't exactly the same. So, instead we use use the
current value of self.num_seen as a non-controversial tie-breaker.