Rename framework benchmarking bases from Time To Bench. PYTHON-2983
It turns out the django benchmarks were not actually broken. What really
happened is that asv released a new version, and so we went from using
v0.2.2 to v0.3. asv v0.3 has some sort of changed benchmark discovery
logic and now detected the TimeWrappingBase, TimeWrappedExecutionBase, and TimeInstrumentBase classes, tried to run them as actual benchmarks, and
died. So, we just need to rename them so they won't get detected anymore.