Change SampledDataSet to an iterable. PYTHON-3032
Our audit logging in serverless mode simply prints out the object prior
to json encoding. This causes our audit logs to print the repr of the
generator object for structures using our SampledDataSet.

In order to fix this problem, SampledDataSet has been converted into an
iterable and application now constructs a list for each data type so
that data is finalized before sending to data_collector.