Store transaction thread_id on enter. PYTHON-3302
Set the thread ID upon entering the transaction. This is done so that
any asyncio tasks will be active and the task ID will be used to store
traces into the trace cache. If thread_id is stored on init, the task
that the transaction will run on may not be active yet, causing the
transaction / traces to be missing when the task activates.