Remove app_id from the required arguments in record_deploy. PYTHON-3453
The interface for record_deploy should not have changed when the payload
format was updated since doing so would introduce a user experience
that's not desirable.

This commit uses the list applications API to retrieve the application
ID based on the application name. As a result, the record_deploy command
no longer requires the app_id.