Improve instrumentation robustness for asyncio. PYTHON-3496
In some of our tests, the loop is retrieved and set prior to the New
Relic instrumentation being loaded.

For example,
https://source.datanerd.us/python-agent/python_agent/blob/f1e2e347c07b4fdd314c488212e26b6e591b2ab5/tests/agent_features/_test_async_coroutine_transaction.py#L5
when _test_async_coroutine_transaction.py is collected by pytest, the
loop is loaded through the asyncio.get_event_loop call. This calls out
to asyncio.set_event_loop prior to instrumentation being loaded
resulting in a test failure when that loop is used later.