Improve support for aiohttp class-based views. PYTHON-3518
The aiohttp_cors library explicitly checks that the handler is a
subclass of web.View

As a result, we must not change the class hierarchy otherwise the
customer can experience a crash.

The class hierarchy is preserved by explicitly wrapping the init
function if one exists. Within aiohttp, the handler class is stored and
the request is handled with a call to `handler(request)` which will
invoke the init in the case of a class based view.