Fix DeprecationWarning in import_hook
For Python >=3.4, using find_loader is deprecated, and importing it
generates a DeprecationWarning. Mirroring wrapt's implementation of a
similar construct in its importer.py (on which this code seems modeled),
we can try to use the new find_spec first,  then the old python 2.7
fallback, as before. Since the newrelic python agent only supports 2.7+
and 3.5+, there is no need to keep support for the deprecated
find_loader at all.

The immediate result of this change is that one less DeprecationWarning
is fired at runtime when using the newrelic python agent. There are
still other DeprecationWarnings originating in this file that are not
fixed by this commit.