Add kafka python consumer (#611)
* Add kafka-python consumer __next__ instrumentation

Co-authored-by: UmaAnnamalai <UmaAnnamalai@users.noreply.github.com>
Co-authored-by: TimPansino <TimPansino@users.noreply.github.com>
Co-authored-by: LallehRafeei <lrafeei@users.noreply.github.com>

* Replace iter(next()) with for loop

The consumer instrumentation of __next__ relies on __next__ getting called until
StopIteration is raised. If this does not happen, the current transaction is not ended.

Because the heartbeat tests were not calling __next__ a second time to end the current
transaction, this caused failures in the consumer tests where the previous
transaction was not getting closed. Because the newrelic application is per session
rather than per function, the unclosed transaction inside the heartbeat tests were
impacting the consumer tests which ran right after.

Co-authored-by: LallehRafeei <lrafeei@users.noreply.github.com>

* Remove duplicated module has attr

Co-authored-by: LallehRafeei <lrafeei@users.noreply.github.com>

Co-authored-by: UmaAnnamalai <UmaAnnamalai@users.noreply.github.com>
Co-authored-by: TimPansino <TimPansino@users.noreply.github.com>
Co-authored-by: LallehRafeei <lrafeei@users.noreply.github.com>