Fix silent failure in validator fixture
Previously, if notice_error did not record an error then the
validate_error_event_attributes_outside_transaction fixture would silently pass. This
is because the validation would not ever run because it was decorated in a stats_engine
notice error block. By moving the validation outside of this decorator, now the
validation will run even if no error is recorded in the stats_engine. This has been
fixed so now the fixture raises if no error is recorded at all.