Fix silent failure in validator fixture (#627)
* Fix silent failure in validator fixture

Previously, if notice_error did not record an error then the
validate_error_event_attributes_outside_transaction fixture would silently pass. This
is because the validation would not ever run because it was decorated in a stats_engine
notice error block. By moving the validation outside of this decorator, now the
validation will run even if no error is recorded in the stats_engine. This has been
fixed so now the fixture raises if no error is recorded at all.

* Fix pylint errors

Fix the following pylint errors:
* tests/testing_support/fixtures.py:2400:46: E0601: Using variable 'backup' before assignment (used-before-assignment)
* tests/testing_support/fixtures.py:2423:37: E0601: Using variable 'backup' before assignment (used-before-assignment)
* tests/testing_support/fixtures.py:2447:59: E0601: Using variable 'original' before assignment (used-before-assignment)
* tests/testing_support/fixtures.py:2452:0: W0102: Dangerous default value [] as argument (dangerous-default-value)