Add threading lock to TraceNode _cache
Since the dictionary is changing sizes during the attempted caching of the dictionary,
this means it is being modified at the same time as it is being used. It is not just a
matter of it changing sizes during a loop so it can't be addressed by a simple caching
of the dictionary before looping). This will require a lock to prevent modification
while it is being used. This adds a threading lock to the TraceNode so that the cache
cannot be updated while it is being looped over (including while the cache is being
cached).