Fix bug in celery where works don't report data (#696)
This fixes Missing information from Celery workers when using MAX_TASKS_PER_CHILD
issue.

Previously, if celery was run with the --loglevel=INFO flag, an agent instance would be
created for the main celery process and after the first worker shutdown, all following
worker's agent instances would point to that agent instance instead of creating a new
instance. This was root caused to incorrectly creating an agent instance when
application activate was not set. Now no agent instance will be created for the main
celery process.