Split below es 8 methods from es 8 methods
Note the previous tests in this file to check whether a method was instrumented, did not test anything because they were checking whether the list
of methods that we instrumented were instrumented instead of whether
there were uninstrumented methods on the es client that we missed.

Because we decided due to lack of reporting of bugs by our customers, to not support the buggy wrapping on previous es versions (below es8), we
only added tests to assert all methods were wrapped from es8+. We also
are only testing es8+ wrapping of methods since the previous versions
wrapping behavior may not have been correct due to the signature of the
methods changing without us detecting it due to lack of tests. Since our customers have not reported any issues, it seems not worth it at this time to go back and fix these bugs.