Remove enable_datastore_instance_feature
This was added in 2016 when the database instance feature was first developed. It
appears to be a method of gating this feature internally within the agent at the time
that it was implemented. However, it is not needed now and database instrumentations
that don't call this are actually broken in that the metric that is used to create the
service map (namely `Datastore/instance/MySQL/<host>/<port>`) does not get created due
to not calling this enable feature function.