Fix NewRelicContextFormatter bug
Previously, there was a bug in NewRelicContextFormatter where "message" was added to the
default set of keys to exclude. This made the default key list (default keys + 1) in
length. Since the logic check was based on the length, rather than presence of certain
keys, anything that had one extra key than the default would not have that extra key
added to the ouput because the length would match the (default keys + 1) length.

The length check is a bit odd here as it leads to a bug in the implementation.
Instead, let's check for presence of keys which is more reflective of our intended
goal than checking the length.