Update attributes in LLM events (#1082)
* Add settings for bedrock

* Add langchain tests

* Add openai v1 tests

* Add openai v0 tests

* Fix NewRelicContextFormatter bug

Previously, there was a bug in NewRelicContextFormatter where "message" was added to the
default set of keys to exclude. This made the default key list (default keys + 1) in
length. Since the logic check was based on the length, rather than presence of certain
keys, anything that had one extra key than the default would not have that extra key
added to the ouput because the length would match the (default keys + 1) length.

The length check is a bit odd here as it leads to a bug in the implementation.
Instead, let's check for presence of keys which is more reflective of our intended
goal than checking the length.

* Remove appName from LLM attrs

* Remove api_key from LLM attrs

* lowercase characters for vendor

* Add new vectorstores

* Fix file path

* Fix merge conflicts

* Fix merge conflicts (again)

* Fix merge conflicts

---------

Co-authored-by: Hannah Stepanek <hstepanek@newrelic.com>
Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>