Attach arn to api_params & request_dict
These are regenerated for each request as opposed to the events obj that persists across
multiple requests. Unfortunately, we have to pass the arn from api_params to the
request_dict because:
1. api_params has validation key checks that exclude unknown params upon generating the
   request_dict
2. the arn is not available after api_params are emitted because emit drops the param
   that contains the arn

So...we capture the arn before it is lost when the api_params are emitted and pop it off
before generating the request dict and put it onto the output request dict to be popped
off later by the make request method and added to the trace to link the entity.