Reorder inheritance on span nodes (#1539)
* Reorder inheritance on span nodes

Core tracing will be adding a bunch of logic to the span_event method which is called
via super on inheriting classes before those inheriting classes add additional
attributes. In order for the core tracing logic to work correctly, this inheritance
call order has to be reversed; the inheriting classes must first add the attributes
and THEN call super.

* Remove *args, **kwargs

* [MegaLinter] Apply linters fixes

---------

Co-authored-by: hmstepanek <30059933+hmstepanek@users.noreply.github.com>